requirements.txt
streamlit>=1.35.0
anthropic>=0.34.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.20.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
